requirements.txt
Appium-Python-Client==4.2.0
selenium==4.25.0
pytest==7.2.1
allure-pytest==2.12.0
adbutils==0.5.0
coloredlogs==10.0
allure-pytest==2.12.0
allure-python-commons~=2.12.0
Faker~=37.0.0
